requirements.txt
discord.py
asyncio
ffmpeg-python
PyNaCl
datetime
git+https://github.com/Rapptz/discord.py